Move the development apps to there own  project and load them in dynamically
