+++ NetDaemon.sln
@@ -28,6 +28,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{32E20E48
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetDaemon.Fakes", "src/Fakes/NetDaemon.Fakes/NetDaemon.Fakes.csproj", "{E1844E4F-2711-4A04-B70C-D67D5D2C8F31}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetDaemon.DevelopmentApps", "src\DevelopmentApps\NetDaemon.DevelopmentApps.csproj", "{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DevelopmentApps", "DevelopmentApps", "{CD3259F6-4CD3-49AA-90D3-3CFEEE0BFF0E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -110,6 +114,18 @@ Global
 		{E1844E4F-2711-4A04-B70C-D67D5D2C8F31}.Release|x64.Build.0 = Release|Any CPU
 		{E1844E4F-2711-4A04-B70C-D67D5D2C8F31}.Release|x86.ActiveCfg = Release|Any CPU
 		{E1844E4F-2711-4A04-B70C-D67D5D2C8F31}.Release|x86.Build.0 = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|x64.Build.0 = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Debug|x86.Build.0 = Debug|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|x64.ActiveCfg = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|x64.Build.0 = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|x86.ActiveCfg = Release|Any CPU
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -121,6 +137,7 @@ Global
 		{E73BF13E-A17F-4B0B-8504-C2A299B83DEE} = {42718370-2C75-4708-8C40-DD59D633C96E}
 		{E79D336A-0BA1-4662-A840-CAB24B3145B9} = {A7F8279D-A148-48B8-8492-BDD96CE1E676}
 		{E1844E4F-2711-4A04-B70C-D67D5D2C8F31} = {32E20E48-2F62-447B-83A2-DB2EA7F61AB3}
+		{E33F4CFC-7D80-48C4-B5FE-107E36CD98E4} = {CD3259F6-4CD3-49AA-90D3-3CFEEE0BFF0E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7C5FBB7F-654C-4CAC-964F-6D71AF3D62F8}
