+++ src/DevelopmentApps/NetDaemon.DevelopmentApps.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <LangVersion>8.0</LangVersion>
+    <Nullable>enable</Nullable>
+ </PropertyGroup>
+  <ItemGroup>
+    <None Update="secrets.yaml">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="apps\HelloWorld\HelloWorld.yaml">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="apps\DebugApp\DebugApp.yaml">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\App\NetDaemon.App\NetDaemon.App.csproj" />
+  </ItemGroup>
+</Project>
