+++ src/DevelopmentApps/apps/DebugApp/DebugApp.yaml
@@ -1,0 +1,2 @@
+debug_app:
+  class: NetDaemon.DevelopmentApps.apps.DebugApp.DebugApp
