+++ src/DevelopmentApps/apps/HelloWorld/HelloWorld.yaml
@@ -1,0 +1,2 @@
+hello_world_app:
+  class: NetDaemon.DevelopmentApps.apps.HelloWorld.HelloWorldApp
