+++ src/DevelopmentApps/secrets.yaml
@@ -1,0 +1,4 @@
+
+# Use this file to store secrets like usernames and passwords.
+# Learn more at https://www.home-assistant.io/docs/configuration/secrets/
+some_password: welcome
