--- src/Service/apps/DebugApp.yaml
@@ -1,2 +1,0 @@
-debug_app:
-  class: Debug.DebugApp
